+++ components/bridge/rpc/api/src/main/java/com/flipperdevices/bridge/rpc/api/model/NameWithHash.kt
@@ -1,0 +1,10 @@
+package com.flipperdevices.bridge.rpc.api.model
+
+import com.flipperdevices.protobuf.storage.Storage
+
+data class NameWithHash(
+    val name: String,
+    val md5: String,
+    val size: Int,
+    val type: Storage.File.FileType
+)