+++ components/bridge/synchronization/impl/src/main/java/com/flipperdevices/bridge/synchronization/impl/repository/flipper/FlipperHashRepository.kt
@@ -1,88 +1,91 @@
 package com.flipperdevices.bridge.synchronization.impl.repository.flipper
 
-import com.flipperdevices.bridge.api.manager.FlipperRequestApi
-import com.flipperdevices.bridge.api.model.FlipperRequestPriority
-import com.flipperdevices.bridge.api.model.wrapToRequest
+import com.flipperdevices.bridge.api.utils.Constants
 import com.flipperdevices.bridge.dao.api.model.FlipperFilePath
+import com.flipperdevices.bridge.dao.api.model.FlipperFileType
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyType
+import com.flipperdevices.bridge.rpc.api.FlipperStorageApi
+import com.flipperdevices.bridge.rpc.api.model.NameWithHash
 import com.flipperdevices.bridge.synchronization.impl.di.TaskGraph
 import com.flipperdevices.bridge.synchronization.impl.model.KeyWithHash
-import com.flipperdevices.bridge.synchronization.impl.model.ResultWithProgress
-import com.flipperdevices.bridge.synchronization.impl.model.trackProgressAndReturn
-import com.flipperdevices.core.ktx.jre.pmap
 import com.flipperdevices.core.log.LogTagProvider
+import com.flipperdevices.core.log.debug
 import com.flipperdevices.core.log.info
 import com.flipperdevices.core.progress.ProgressWrapperTracker
-import com.flipperdevices.protobuf.main
-import com.flipperdevices.protobuf.storage.md5sumRequest
+import com.flipperdevices.protobuf.storage.Storage
 import com.squareup.anvil.annotations.ContributesBinding
-import kotlinx.coroutines.flow.callbackFlow
-import kotlinx.coroutines.flow.single
-import java.util.concurrent.atomic.AtomicLong
+import java.io.File
 import javax.inject.Inject
+
+const val SIZE_BYTES_LIMIT = 10 * 1024 * 1024 // 10MiB
 
 interface FlipperHashRepository {
     suspend fun getHashesForType(
-        flipperKeyType: FlipperKeyType,
+        keyType: FlipperKeyType,
         tracker: ProgressWrapperTracker
     ): List<KeyWithHash>
 }
 
 @ContributesBinding(TaskGraph::class, FlipperHashRepository::class)
 class FlipperHashRepositoryImpl @Inject constructor(
-    private val requestApi: FlipperRequestApi,
-    private val keysListingRepository: KeysListingRepository
+    private val flipperStorageApi: FlipperStorageApi
 ) : FlipperHashRepository, LogTagProvider {
     override val TAG = "HashRepository"
 
     override suspend fun getHashesForType(
-        flipperKeyType: FlipperKeyType,
+        keyType: FlipperKeyType,
         tracker: ProgressWrapperTracker
     ): List<KeyWithHash> {
-        val flipperKeys = keysListingRepository.getKeysForType(flipperKeyType)
-        tracker.onProgress(current = 0.3f)
-        val hashProgressTracker = ProgressWrapperTracker(
-            min = 0.3f,
-            max = 1f,
-            progressListener = tracker
-        )
-
-        return calculateHash(flipperKeys).trackProgressAndReturn {
-            // Progress
-            hashProgressTracker.report(it.currentPosition, it.maxPosition)
+        val fileTypePath = File(Constants.KEYS_DEFAULT_STORAGE, keyType.flipperDir).path
+        val files = flipperStorageApi.listingDirectoryWithMd5(fileTypePath)
+        info { "Receive ${files.size} files" }
+        tracker.onProgress(1f)
+        return files.filter { isValidFile(it, keyType) }.map {
+            KeyWithHash(keyPath = FlipperFilePath(keyType.flipperDir, it.name), hash = it.md5)
         }
     }
 
-    private fun calculateHash(
-        keys: List<FlipperFilePath>
-    ) = callbackFlow {
-        val alreadyHashReceiveCounter = AtomicLong(0)
-        info { "Start request hashes for ${keys.size} keys" }
-
-        val hashList = keys.pmap { keyPath ->
-            val hash = receiveHash(keyPath)
-            send(
-                ResultWithProgress.InProgress(
-                    currentPosition = alreadyHashReceiveCounter.incrementAndGet(),
-                    maxPosition = keys.size.toLong()
-                )
-            )
-            return@pmap KeyWithHash(keyPath, hash)
+    private fun isValidFile(nameWithHash: NameWithHash, requestedType: FlipperKeyType): Boolean {
+        if (nameWithHash.type != Storage.File.FileType.FILE) {
+            debug {
+                "File ${nameWithHash.name} is not file. This is folder. Ignore it"
+            }
+            return false
         }
+        if (nameWithHash.size > SIZE_BYTES_LIMIT) {
+            debug {
+                "File ${nameWithHash.name} skip, because current size limit is $SIZE_BYTES_LIMIT, " +
+                    "but file size is ${nameWithHash.size}"
+            }
+            return false
+        }
+        if (nameWithHash.name.startsWith(".")) {
+            debug {
+                "File ${nameWithHash.name} skip, because it starts with dot"
+            }
+            return false
+        }
+        val extension = nameWithHash.name.substringAfterLast(".")
 
-        send(ResultWithProgress.Completed(hashList))
-        close()
-    }
-
-    private suspend fun receiveHash(
-        keyPath: FlipperFilePath
-    ): String {
-        return requestApi.request(
-            main {
-                storageMd5SumRequest = md5sumRequest {
-                    path = keyPath.getPathOnFlipper()
-                }
-            }.wrapToRequest(FlipperRequestPriority.BACKGROUND)
-        ).single().storageMd5SumResponse.md5Sum
+        if (FlipperFileType.getByExtension(extension) == FlipperFileType.SHADOW_NFC &&
+            requestedType == FlipperKeyType.NFC
+        ) {
+            return true
+        }
+        val fileTypeByExtension = FlipperKeyType.getByExtension(extension)
+        if (fileTypeByExtension == null) {
+            debug {
+                "File ${nameWithHash.name} skip, because we don't support this file extension ($extension)"
+            }
+            return false
+        }
+        if (fileTypeByExtension != requestedType) {
+            debug {
+                "File ${nameWithHash.name} skip, because folder type ($requestedType) " +
+                    "and extension type ($fileTypeByExtension) is not equals"
+            }
+            return false
+        }
+        return true
     }
 }